requirements.txt
# Runtime
python-dotenv>=0.21.0
fastapi>=0.95.0
uvicorn[standard]>=0.22.0     # ASGI server

# Validation
pydantic>=1.10.0

# Database / async
asyncpg>=0.27.0
sqlalchemy>=1.4.0            # SQLAlchemy with asyncio support
elasticsearch>=8.0.0        # Elasticsearch client

# HTTP clients
httpx>=0.24.0                # async http client

# AI / LLM
google-genai

# Type helpers
typing_extensions>=4.5.0

# Frontend
streamlit

# Vector operations
numpy>=1.21.0
